+++ src/stories/select.stories.tsx
@@ -31,61 +31,61 @@ stories.add("Default", () => (
 stories.add("General Usage", () => {
   const [value, setValue] = React.useState("2");
 
+  const handleSelect = (e: any) => {
+    setValue(e.target.value);
+  };
+
+  const handleButton = () => {
+    setValue(value === "1" ? "2" : "1");
+  };
+
   return (
-    <div>
-      <Select items={items} valueKey="id" value={value} leftIcon={"close"} onChange={(e) => setValue(e.target.value)}/>
-      <br />
-      <br />
-      <Select items={items} valueKey="id" rightIcon={"chevron-down"} value="2" />
-      <br />
-      <br />
-      <Select items={items} valueKey="id" leftIcon="filter" leftIconColor="primary" rightIcon="close" value="3" />
-      <br />
-      <br />
-      <div className={"flex"}>
-        <Select className={"select-w40"} items={items} fluid valueKey="id" leftIcon="filter" leftIconColor="primary" rightIcon="close" value="3" />
-        <Button className={"button-w60"} onClick={(e) => {setValue(value === "1" ? "2" : "1"); }} fluid>Hello World</Button>
-      </div>
+    <div className={"flex"}>
+      <Select
+        className={"select-w40"}
+        items={items}
+        fluid
+        valueKey="id"
+        leftIcon="filter"
+        leftIconColor="primary"
+        rightIcon="close"
+        value={value}
+        onChange={handleSelect}
+      />
+      <Button className={"button-w60"} onClick={handleButton} fluid>Hello World</Button>
     </div>
   );
 });
 
 stories.add("Icon", () => {
-  const [value, setValue] = React.useState("2");
 
   return (
     <div>
-      <Select items={items} valueKey="id" value={value} leftIcon={"close"} onChange={(e) => setValue(e.target.value)}/>
+      <Select items={items} valueKey="id" leftIcon={"close"} />
       <br />
       <br />
-      <Select items={items} valueKey="id" rightIcon={"chevron-down"} value="2" />
+      <Select items={items} valueKey="id" rightIcon={"chevron-down"} />
       <br />
       <br />
-      <Select items={items} valueKey="id" leftIcon="filter" leftIconColor="primary" rightIcon="close" value="3" />
-      <br />
-      <br />
-      <div className={"flex"}>
-        <Select className={"select-w40"} items={items} fluid valueKey="id" leftIcon="filter" leftIconColor="primary" rightIcon="close" value="3" />
-        <Button className={"button-w60"} onClick={(e) => {setValue(value === "1" ? "2" : "1"); }} fluid>Hello World</Button>
-      </div>
+      <Select items={items} valueKey="id" leftIcon="filter" leftIconColor="primary" rightIcon="close" />
     </div>
   );
 });
 
 stories.add("Fluid", () => (
   <div>
-    <Select fluid items={items} rightIcon="close" selected="3" />
+    <Select fluid items={items} rightIcon="close" />
   </div>
 ));
 
 stories.add("Disabled", () => (
   <div>
-    <Select disabled items={items} rightIcon="close" selected="3" />
+    <Select disabled items={items} rightIcon="close" />
   </div>
 ));
 
 stories.add("Error", () => (
   <div>
-    <Select error items={items} rightIcon={"close"} selected={"3"} />
+    <Select error items={items} rightIcon={"close"} />
   </div>
 ));